+++ Models/Request/Satuan.cs
@@ -6,7 +6,7 @@ namespace Dotnet5api.Models
     public class Satuan
     {
         [Key] 
-        public int id_unit { get; set; }
+        public long id_unit { get; set; }
         public string kode_unit { get; set; }
         public string nama_unit { get; set; }
         public string note_unit{ get; set; }
